Allow link check to pass if we time out
If we time out, the remote server is probably down and it's not our fault.
Before this commit, the check would fail, preventing us from deploying the
site. Now, the check passes and just prints a warning, still exiting with exit
code 0 if there are no other errors.

note: this relies on curl printing HTTP code 000 on timeout, and there may be
some other reasons for 000.