fix(code): Use same colors for all code blocks
Before this commit, code blocks with syntax highlighting (so with a language
specified after the leading triple backticks) would have a different background
color than code blocks without syntax highlighting.
To fix this, the selector `.chroma` was changed to `pre, code` in code.sass, so
that the styles in there now apply to all code blocks (not just those with
syntax highlighting). The styles in elements.sass were also updated to apply
background color only to inline code blocks (to have enough contrast with the
background without adding a border).