Remove extra GPG key
I'm not sure why this key is here.
When importing the key with GPG, it looks like it's my tarneo@tarneo.fr
key, but the WKD hash (the filename) would be for admin@tarneo.fr, which
is not a key I have.
There is no security implication here: it looks like the key was just
copied from the other one.